refactor(ai): migrate list_job_postings to tools registry
