+++ src/lib/ai/tools/registry/index.ts
@@ -5,6 +5,7 @@ import { listMembersModule } from "./list-members";
 import { listEventsModule } from "./list-events";
 import { listAnnouncementsModule } from "./list-announcements";
 import { listDiscussionsModule } from "./list-discussions";
+import { listJobPostingsModule } from "./list-job-postings";
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 const MODULES: ReadonlyArray<ToolModule<any>> = [
@@ -12,6 +13,7 @@ const MODULES: ReadonlyArray<ToolModule<any>> = [
   listEventsModule,
   listAnnouncementsModule,
   listDiscussionsModule,
+  listJobPostingsModule,
 ];
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
